% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wikipedia_references.R
\name{clean_wikipedia_references}
\alias{clean_wikipedia_references}
\title{Clean Wikipedia References}
\usage{
clean_wikipedia_references(references)
}
\arguments{
\item{references}{A character vector of references.}
}
\value{
Null but a message is printed to the console with cleaned references.
}
\description{
This function cleans a character vector of references by removing
unnecessary characters such as quotation marks and apostrophes.
}
