% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikipedia_references.R
\name{get_wikipedia_references}
\alias{get_wikipedia_references}
\title{Get Wikipedia References}
\usage{
get_wikipedia_references(url)
}
\arguments{
\item{url}{The URL of the Wikipedia page.}
}
\value{
A character vector of formatted references.
}
\description{
This function scrapes and displays references from a valid Wikipedia page.
}
