% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikipedia_tables.R
\name{get_wikipedia_tables}
\alias{get_wikipedia_tables}
\title{Scrape Wikipedia Tables and Choose One}
\usage{
get_wikipedia_tables(url)
}
\arguments{
\item{url}{The URL of the Wikipedia page.}
}
\value{
A dataframe representing the selected table, or NULL if no table
was selected.
}
\description{
This function scrapes all tables with captions from a Wikipedia page
specified by its URL, lists them, and allows the user to select one.
}
