% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_systems}
\alias{metro_systems}
\title{A Wikipedia page of metro systems}
\format{A list of HMTL nodes}
\source{
Wikipedia. List of Metro Systems. \url{https://en.wikipedia.org/wiki/List_of_metro_systems}
}
\usage{
metro_systems
}
\description{
A Wikipedia page of metro systems
}
\examples{
\dontrun{
 metro_systems
}
}
\keyword{datasets}
