% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny}
\alias{ny}
\title{The wikipedia page for New York City}
\format{A list of HMTL nodes}
\source{
Wikipedia. New York City. \url{https://en.wikipedia.org/wiki/New_York_City}
}
\usage{
ny
}
\description{
The wikipedia page for New York City
}
\examples{
\dontrun{
 ny
}
}
\keyword{datasets}
