% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_wikipedia_references.R
\name{save_wikipedia_references}
\alias{save_wikipedia_references}
\title{Save cleaned references to a text file.}
\usage{
save_wikipedia_references(references, output_file)
}
\arguments{
\item{references}{A character vector of cleaned references.}

\item{output_file}{The path to the output text file.}
}
\value{
NULL, but the cleaned references are saved to the specified file.
}
\description{
This function saves the cleaned references to a specified text file.
}
