% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_wikipedia_table.R
\name{save_wikipedia_table}
\alias{save_wikipedia_table}
\title{Save a Data Frame to Various Formats}
\usage{
save_wikipedia_table(data_frame, file_name, format)
}
\arguments{
\item{data_frame}{The dataframe to be saved.}

\item{file_name}{The name of the output file (without the file extension).}

\item{format}{The desired output format ('csv', 'xlsx', 'tsv', or 'json').}
}
\value{
NULL (file is saved to the specified location).
}
\description{
This function saves a dataframe to different file formats, including CSV,
Excel (XLSX), TSV (Tab-Separated Values), and JSON. More formats may be
supported in the future.
}
