% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_wikipedia_article.R
\name{summarize_wikipedia_article}
\alias{summarize_wikipedia_article}
\title{Summarize Wikipedia Article}
\usage{
summarize_wikipedia_article(url)
}
\arguments{
\item{url}{The URL of the Wikipedia page.}
}
\value{
A character vector containing the summary of the Wikipedia page.
}
\description{
This function retrieves a summary of a Wikipedia page from an API
using the article's URL.
}
