% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_names.R
\name{wiki_names}
\alias{wiki_names}
\title{Tidy column names in tables}
\usage{
wiki_names(data, replace_all = c(""), remove = NULL, rename = NULL,
  lowercase = TRUE)
}
\arguments{
\item{data}{The data frame or tibble to tidy.}

\item{remove}{A regular expression or vector of regular expressions to remove in all header names. For example,
the argument c("aa","b") would delete ALL occurances of the characters "aa" and "b".}

\item{rename}{A vector of the form c("old_name1", "new_name1", "old_name2", "new_name2"). In this case, "old_name1"
or "old_name2" could be replaced with an integer indicating the column index in which the new name should be inserted.}

\item{lowercase}{A boolean indicating whether the output names should be forced to lowercase. Defaults to TRUE.}

\item{replace}{A vector of regular expressions and character string. The first element in the vector is
replaced by the second, the third is replaced by the fourth, and so on. If a vector with an odd length is
passed, the last element will be ignored
will replace the first value with the second. In other words c("foo","bar") in the title
would replace all occurances of "foo" with "bar" in the new menu.}
}
\description{
A helper function to clean up column names.
}
\note{
Order of Operations: 1) replace punctuation and spaces with underscores, 2) rename,
    3) replace_all, 4) remove, 5) to lowercase. This could impact the results of the function.
}
\examples{
wiki_table("List_of_metro_systems") \%>\% wiki_names()

}
