% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_table.R
\name{wiki_table_format}
\alias{wiki_table_format}
\title{Parse the format of an HTML table.}
\usage{
wiki_table_format(table, skip = "auto", header_length = "auto")
}
\arguments{
\item{table}{An HTML table object.}

\item{skip}{The number of rows to skip before collecting data. This is useful for omitting
full-width "title" cells. Takes an integery and defaults to 0.}

\item{header_length}{Set to a number greater than one to deal with multi-row headers.
Takes an integer and defaults to 1.}
}
\value{
Returns a list object that contains information about the format of \code{table}.
}
\description{
A function to parse an HTML table's format.
}
\examples{
wiki_table("https://wikipedia.org/wiki/List_of_metro_systems")
wiki_table("List_of_metro_systems")

}
