% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_table.R
\name{wiki_table_header}
\alias{wiki_table_header}
\title{A Function to Parse an HTML Table Header}
\usage{
wiki_table_header(table, header_length = "auto", skip = "auto",
  col_names = NULL, rm_text = NULL, rm_brackets = TRUE,
  rm_parens = FALSE)
}
\arguments{
\item{table}{An HTML table object.}

\item{header_length}{Set to a number greater than one to deal with multi-row headers.
Takes an integer and defaults to 1.}

\item{skip}{The number of rows to skip before collecting data. This is useful for omitting
full-width "title" cells. Takes a character vector of rows to skip or the string
 \code{"auto"}, and defaults to \code{"auto"}. If argument is set to \code{"auto"},
 parser will automatically remove cells that take up a full row.}

\item{exclude_brackets}{Whether to exclude brackets and their contents in output.
Takes a boolean and defaults to TRUE.}

\item{exclude_parens}{Whether to exclude parenthesis and their contents in output.
Takes a boolean and defaults to FALSE.}
}
\value{
Returns a character vector of parsed headers.
}
\description{
Takes an HTML table, parses it's format, and returns a character
    vector as long as the input table is wide.
}
\examples{
wiki_table("https://wikipedia.org/wiki/List_of_metro_systems")
wiki_table("List_of_metro_systems")

}
