% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_card.R
\name{ws_get_card}
\alias{ws_get_card}
\title{Get Data from Wikipedia Card}
\usage{
ws_get_card(page, format = "long")
}
\arguments{
\item{format}{Either 'long' or 'wide'. 'long' returns an output with two columns (header and data), 'wide' returns an output with a column for each data entry.}

\item{html}{Either a url to a wikipediea page, or an object that contains the body of a wikipedia page (e.g. from ws_scrape_page).}
}
\description{
A function to extract data from the "infocards" some Wikipedia pages provide.
}
\examples{
ws_get_card("wiki/New_York_City")

# OR
# get page THEN get card

page <- ws_get_page("New_York_City") # get page
ws_get_card(page) # then get data from the card

}
