% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_geometry.R
\name{ws_get_geometry}
\alias{ws_get_geometry}
\title{Get Lat/Long Coordinates from Wikipedia Page}
\usage{
ws_get_geometry(data, delete_coords = T, coordinates)
}
\arguments{
\item{data}{A data frame that contains a column of coordinates from Wikipedia.}

\item{delete_coords}{Boolean value that indicates whether to remove the original
coordinates column.}

\item{coordinates}{A character object that contains the name of the column
that contains the coordiantes to parse. If missing, the function will look for
columns named "coords" or "coordinates".}
}
\value{
Returns a the dataframe passed in the argument data, with columns "lat" and "lon"
    appended. By default, the original "coordinates" column is deleted.
}
\description{
A helper function to parse Wikipedia-style geographic coordinates.
}
\examples{
ws_get_table("List_of_metro_systems") \%>\% ws_tidy_names()

}
