% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_section.R
\name{ws_get_section}
\alias{ws_get_section}
\title{Get a Section from a Wikipedia Page}
\usage{
ws_get_section(page, section)
}
\arguments{
\item{page}{Either a url to a wikipediea page, or an object that contains the body of a wikipedia page (e.g. from ws_scrape_page).}

\item{section}{The header or css id of the section to retrieve}
}
\description{
Returns an HTML object that contains the requested section of the web page.
}
\examples{
ws_get_page("List_of_metro_systems")
# is equivelant to
ws_get_page("https://en.wikipedia.org/wiki/List_of_metro_systems")

}
