% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_table.R
\name{ws_get_table}
\alias{ws_get_table}
\title{Scrape a Table from Wikipedia Page}
\usage{
ws_get_table(page, table_num = 1, skip = 0, header_length = 1,
  exclude_brackets = TRUE, exclude_parens = FALSE, format = NULL)
}
\arguments{
\item{page}{Either the url of a Wikipedia, or an object that contains a Wikipedia page}

\item{table_num}{The index of the table to get data from, defaults to 1 (first table)}

\item{skip}{The number of rows to skip before collecting data. This is useful for omitting full-width "title" cells.
Takes an integery and defaults to 0.}

\item{header_length}{set to a number greater than one to deal with multi-row headers. Takes an integer and defaults to 1.}

\item{exclude_brackets}{Whether to exclude brackets and their contents in output. Takes a boolean and defaults to TRUE.}

\item{exclude_parens}{Whether to exclude parenthesis and their contents in output. Takes a boolean and defaults to FALSE}
}
\value{
data_frame
}
\description{
A function to extract a table from a Wikipedia page.
}
\examples{
ws_get_table("https://en.wikipedia.org/wiki/List_of_metro_systems")

}
