% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_table.R
\name{wiki_table}
\alias{wiki_table}
\title{Scrape a Table from Wikipedia Page}
\usage{
wiki_table(page, table = 1, skip = 0, header_length = 1,
  col_names = NULL, exclude_brackets = TRUE, exclude_parens = FALSE,
  format = NULL, delay = 1)
}
\arguments{
\item{page}{Either the url of a Wikipedia, or an object that contains a Wikipedia page}

\item{table}{An integer that specifies which table to get data from. Defaults to 1,
which retrieves the first table element in the HTML object or web page passed.}

\item{skip}{The number of rows to skip before collecting data. This is useful for omitting
full-width "title" cells. Takes an integery and defaults to 0.}

\item{header_length}{Set to a number greater than one to deal with multi-row headers.
Takes an integer and defaults to 1.}

\item{col_names}{Optional argument that takes a character vector to name columns in
the output table.}

\item{exclude_brackets}{Whether to exclude brackets and their contents in output.
Takes a boolean and defaults to TRUE.}

\item{exclude_parens}{Whether to exclude parenthesis and their contents in output.
Takes a boolean and defaults to FALSE}

\item{delay}{Rate at which to throttle calls. There is no delay if the function is passed an HTML object
(e.g. from wiki_page). Defaults to 1, can be turned off by setting
to 0. Time between calls is determined by multiplying the value of this parameter with
the response time by the server.}
}
\value{
Returns a dataframe (tibble) that contains the data from the table
    specified by the table argument.
}
\description{
A function to extract a table from a Wikipedia page.
}
\examples{
wiki_table("https://wikipedia.org/wiki/List_of_metro_systems")
wiki_table("List_of_metro_systems")

}
