% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_page.R
\name{ws_get_page}
\alias{ws_get_page}
\title{Get HTML from Wikipedia Page}
\usage{
ws_get_page(page, url = "https://en.wikipedia.org/wiki/")
}
\arguments{
\item{page}{Extension of the page, e.g. "New_York_City"}

\item{url}{The base url of the site to visit, defaults to "https://en.wikipedia.org/wiki/"}
}
\description{
A helper function to extract a table from a Wikipedia page.
}
\examples{
ws_get_page("New_York_City")
# is equivelant to
ws_get_page("https://en.wikipedia.org/wiki/New_York_City")

}
